Classpathresourcetemp (#1122)
* auto generate minibatches that are balanced

* add create methods for javacpp pointers

* add standard interface for data normalization

* add load from input stream

* add write/read for dataset on streams

* reduce gc on shape/strideof views on ndarray

* add classpath resource abilty to get a temp file from an archive

* close input stream

* local

+++ nd4j-common/src/main/java/org/nd4j/linalg/io/ClassPathResource.java
@@ -48,7 +48,7 @@ public class ClassPathResource extends AbstractFileResolvingResource {
     }
 
     public final ClassLoader getClassLoader() {
-        return this.classLoader != null?this.classLoader:this.clazz.getClassLoader();
+        return this.classLoader != null ? this.classLoader:this.clazz.getClassLoader();
     }
 
     /**
@@ -61,13 +61,21 @@ public class ClassPathResource extends AbstractFileResolvingResource {
      */
     public File getTempFileFromArchive() throws IOException {
         InputStream is = getInputStream();
+        //directory
         File tmpFile = new File(path + "tmp");
-        BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(tmpFile));
-        tmpFile.deleteOnExit();
-        IOUtils.copy(is,bos);
-        bos.flush();
-        bos.close();
-        is.close();
+        if(is != null) {
+            BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(tmpFile));
+            tmpFile.deleteOnExit();
+            IOUtils.copy(is,bos);
+            bos.flush();
+            bos.close();
+            is.close();
+        }
+        //extract
+        else {
+
+        }
+
         return tmpFile;
 
     }
